requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Camera and Computer Vision
opencv-python-headless==4.10.0.84
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.3

# Camera and Computer Vision
opencv-python-headless==4.10.0.84
flask==3.0.0
flask-cors==4.0.0
numpy==1.26.3

# AI and ML
google-genai>=0.5.0
google-generativeai>=0.6.0
google-api-core>=2.15.0
elevenlabs==0.2.27
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0

# Audio Pipeline - REQUIRED
faster-whisper>=1.0.3
# Audio Pipeline
faster-whisper>=1.0.3
pyaudio==0.2.14
webrtcvad==2.0.10
pydub==0.25.1

# Computer Vision (for camera tracking)
opencv-python==4.9.0.80
mediapipe==0.10.14

# WebSocket (for hand tracking streaming)
websockets>=13.0,<15.1

# Web (for UI)
flask==3.0.0
flask-cors==4.0.0

# Utilities
loguru==0.7.2
colorama==0.4.6

# Testing
pytest==7.4.4

# Hardware Control - Arduino Serial Communication
pyserial==3.5

# Authentication & Database
bcrypt==4.1.2
pyjwt==2.8.0
pymongo==4.6.1
cryptography==42.0.0
